perbaikan table user&shipping_address

+++ database/migrations/2025_04_12_171033_add_quantity_to_carts_table.php
@@ -12,14 +12,21 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('carts', function (Blueprint $table) {
-            $table->integer('quantity')->default(1);
+            if (!Schema::hasColumn('carts', 'quantity')) {
+                $table->integer('quantity')->default(1);
+            }
         });
     }
 
+    /**
+     * Reverse the migrations.
+     */
     public function down(): void
     {
         Schema::table('carts', function (Blueprint $table) {
-            $table->dropColumn('quantity');
+            if (Schema::hasColumn('carts', 'quantity')) {
+                $table->dropColumn('quantity');
+            }
         });
     }
 };
